e2e: move log file from test to app package
As the log creation binds more to application, moving log from test
package to app package.
